Split DeviceClient and ModuleClient
